refactor : userDTO, dealerDTO 병합
